+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+[1.3.0] 15.04.2026
+
+- [FAM-170] Category delete
+- [FAM-162] Create/edit category/subcategory
+- [FAM-163] Categires screen
+- [FAM-164] Category create/edit screen design
+- [FAM-161] Category screen design
+
 [1.2.0] 10.04.2026
 
 - [FAM-160] Categori esxpense